adapted to moved AbstractWorkflowConfigurationBuilder
