+++ bundles/de.uka.ipd.sdq.dsexplore/src/de/uka/ipd/sdq/dsexplore/launch/DSELaunch.java
@@ -14,7 +14,7 @@ import org.palladiosimulator.analyzer.workflow.core.configurations.PCMWorkflowCo
 import de.uka.ipd.sdq.workflow.WorkflowExceptionHandler;
 import de.uka.ipd.sdq.workflow.WorkflowProcess;
 import de.uka.ipd.sdq.workflow.jobs.IJob;
-import de.uka.ipd.sdq.workflow.launchconfig.core.AbstractWorkflowConfigurationBuilder;
+import de.uka.ipd.sdq.workflow.launchconfig.core.configbuilder.AbstractWorkflowConfigurationBuilder;
 import de.uka.ipd.sdq.workflow.logging.console.LoggerAppenderStruct;
 import de.uka.ipd.sdq.workflow.ui.UIBasedWorkflowExceptionHandler;
